Switch to unsigned frequency in all places
Accidentally had size_t in create but not in popup. Using u32 could produce a slightly smaller db

+++ include/db.h
@@ -13,7 +13,7 @@ typedef struct database_s database_t;
 typedef struct {
     s8 word;
     s8 reading;
-    size_t frequency;
+    u32 frequency;
 } frequency_entry;
 
 database_t *_nonnull_ db_open(char *dbpath, bool readonly);
